requirements.txt
openpyxl==3.1.5
pandas==2.2.3
python-snap7==2.0.2
paho-mqtt==2.1.0
pyarrow==19.0.1
opcua
cryptography==45.0.2
